Refactor the code sparate the video and audio page
